+++ function/faculty/delete.js
@@ -2,7 +2,12 @@ const FacultyModel = require('../../model/facultyModel');
 
 const deleteFaculty = async (req, res) => {
   try {
-    const id = req.params.id; // Access the ID from the request parameters
+    const {id} = req.params;
+
+    if(!id){
+      return res.status(404).json({ message: "Invalid request!" });
+    }
+
     const facultyData = await FacultyModel.deleteOne({ _id: id });
 
     if (facultyData.deletedCount === 1) {
@@ -10,6 +15,7 @@ const deleteFaculty = async (req, res) => {
     } else {
       return res.status(404).json({ message: 'Faculty not found' });
     }
+    
   } catch (error) {
     console.error(error);
     return res.status(500).json({ message: 'Internal Server Error' });
