+++ function/semister/delete.js
@@ -2,7 +2,12 @@ const SemisterModel = require('../../model/semisterModel');
 
 const deleteSemister = async (req, res) => {
   try {
-    const id = req.params.id; // Access the ID from the request parameters
+    const {id} = req.params;
+
+    if(!id) {
+      return res.status(404).json({ message: 'Invalid request!' });
+    }
+
     const semisterData = await SemisterModel.deleteOne({ _id: id });
 
     if (semisterData.deletedCount === 1) {
@@ -10,6 +15,7 @@ const deleteSemister = async (req, res) => {
     } else {
       return res.status(404).json({ message: 'Semester not found' });
     }
+
   } catch (error) {
     console.error(error);
     return res.status(500).json({ message: 'Internal Server Error' });
